Were many things wrong with the release workflow
Main issue was that you can't get outputs from one job into the other,
you can only get outputs from one step to the other in the same job. The
create release action generates a upload_url output that can then be
used by the upload release asset action. This upload_url needs to be
passed along in an artifact to then be read in a job that can be
parameterized by the OS e.g. ubuntu and macos.

Also only zip the binary so it can be easily unzipped to /usr/bin or
similar.