Supporting code for deploying the cloud
- Adds an optional value to values.yaml file telling the controller.yaml
manifest where it can find the password secret. Feel like this is a good
balance with not putting the whole db address in a secret while keeping
the password secret.
- Modifies the database backend constructors and related bits to take
URLs, I think we should be using URLs as much internally as possible but
this is in support of the above change.
- Adds a sql.Open call to the postgres backend so there is a bit more to
test with
- go.mod, go.sum file changes are from running `go mod tidy`.