Adds the ability for tilt to run tests
To support this I split up the containers so that now there are 4
containers. A base container that downloads the dependencies and builds
the executable, a test container that builds off the base that just runs
`make test`, a controller container that runs `./privacyai controller
start`, and a connector container that runs `./privacyai connector
start`.

This also includes an "test" integration test to make sure integration
tests with the database will actually work in the future. It just makes
a simple query to the postgres backend.