Introduce dbtest package for creating test dbs
This commit introduces the `database/dbtest` package which contains
a framework for creating and tearing down databases to be used inside of
integration test suites.

This package continues the interface design of the `database` by
offering a `TestDatabase` interface which backends must conform too. As
a first backend, it adds support for Postgres.

Tests are now easier to write, an example:

```golang
func TestAThing(t *testing.T) {
    gm.RegisterTestingT(t)

    ctx := context.Background()

    testDB, err := dbtest.New(os.Getenv("DB_URL"))
    gm.Expect(err).To(gm.BeNil())
    defer testDB.Teardown(ctx)

    err = testDB.Setup()
    gm.Expect(err).To(gm.BeNil())

    t.Run("test a thing", func(t *testing.T) {
        // do stuff
    })

    t.Run("woo", func(t *testing.T) {

    })

    t.Run("haha", func(t *testing.T) {

    })
}
```