Add singleton support for test databases
Instead of creating a new test database for every single test block we
now support a Wrapper over test databases that only creates a database
once and tears down once across test runs.

The creation of the wrapper is managed by the `dbtest.New()` method. It
also maintains a registry of wrapped databases. This can be called as a
part of test setup, teardown, etc., across tests.