Switch to jackc/pgx from lib/pq
The pg driver created by jackc has several optimizations implemented,
offers pooling out of the box, and also provides native support for many
Go types.

At the same time, I've introduced the concept of a "service id" which
can be passed into an instance of a Connector or Controller to name a
specific instance. This value can be set via the `CAPE_SERVICE_ID`
environment variable.

When the controller connects to postgres it passes this value as a
runtime parameter allowing for an operator to query postgres for this
value.

Check it out: https://github.com/jackc/pgx