Namespace all environment variables
It's possible that more than one service could be running in the same
process space that clash on the same environment variable (e.g. PORT,
HOST, etc). Furthermore, it's common to sometimes set these defaults and
forget about them.

To prevent that from happening, a best practice is to namespace your
environment variables. To that end, I've namespaced all of our
environment variables as `CAPE_{{thing}}`.