Introduce c-tor for Primitive types
For all basic primitive types I've introduced constructors that make it
_much_ simpler to create primitive types.

This drastically reduces the amount of boilerplate required to implement
a Primitive while also giving nice convenience functions for creating
primitives.

I've also marked Role & Policy as mutable as a user may want to change
the name of a role or policy in future without having to create new
ones.