Don't auto create a pg deployment every time you run `tilt up`
This was causing some problems due to the nature of the helm chart
we were using but it is also kind of a weird thing to try and do
anyway.

Not sure this is the best forever solution but should get us pretty far:

There are now 2 commands added to tilt which can be invoked MANUALLY (and only manually)
One of these will create a DB, and the other will delete it.