Primitives _must_ implement GetType
If a primitive does not implement GetType then when you initialize an
empty Primitive you must supply the type value. By changing this
behaviour such that every Entity must implement `GetType()` we can now
determine the type of the entity through the value returned by the
functioning call.

This makes working with the database much simpler.