Implement QueryOne
In this commit, I've introduced the ability to Query a specific
entity from the database.

The tests were randomly failing because maps don't guarantee the key
ordering during iteration.

To fix this, I've added a sort prior to building the statement. This is
_not_ ideal but at a minimum fixes the tests. We will probably want to
profile this code in future and then rebuild it for speed.

That said, it's not a critical area of our code base -- we won't be
building large queries so the overall impact of the additional sort
should be low.