Implement Primities & Add Constraints
This commit ensures that constraints exist for each value that must be
pulled off of the `data` column into other columns inside the row. These
are important to ensure that the foreign references are setup
appropriately.

It also ensures that the `identity_id` column cannot be null for the
`tokens` table.

Closes dropoutlabs/planning#772