Rearrange Primitives & Database pkg
This commit rearranges the primitives and database package by pushing
`types`, `id`, and several testing facilities into the database package
and out of the `primitives` package.

The general idea behind this change is to remove the dependency between
database on `primitives` and turn `primitives` into a package that uses
the types presented by the `database` package.

This does lead to a strong coupling of the `database` package to things
that rely on `primitives` as they need both. However, this helps to pave
a road where we could extract the `database` package into an external
library.