Introduce Logging Middleware & Flags
In this commit we introduce our logging framework based on `zerolog`
which is a very fast and efficient logger. To do this, we also needed to
introduce the concept of middleware to handle panics that occur as a
part of a request's call stack and to assign an identifier to the
request/return it on the response as a header.

Finally, in order to surface control of the logger, we've introduced two
flags that can be used for setting the logger type and associated
logging level.

Closes dropoutlabs/planning#778