Convert 'CAPE_DB_PASSWORD' to a standalone env var
With this commit, we've remove _every single_ password that comes in via
a flag -- they're all supported via environment variables now!

At the same time, I fixed a bug where we were appending the password as
a query parameter instead of setting it as the password in the userinfo
section of the uri.