Introduce Harness & framework.Server
In order to introduce the usage of `httptest.Server` I've had to
abstract away the "creation" of the Controller handler from the actual
server runtime.

To do this, I've introduced the concept of `framework.Server` which is
responsible for the network/transport layer servering and refactor
`controller.Controller` to return a `http.Handler`. The command line
uses `framework.Server` to start the server.

For integration testing, I've introduced the
`controller.harness.Harness` type which manages the dependencies
required for setting up & running a controller locally for end-to-end
integration testing. This new type replaces `controller.TestController.

In addition, I've introduced the concept of a `Manager` to the `harness`
package which is responsible for managing "application" state such as
whether or not the controller has been "setup" and any additional state
that we want to track to make it easier for us to write integration
tests.

Closes dropoutlabs/planning#815