Introduce auth.Keypair
In preparation for enabling the controller to store the data used to
generate its public and private keypair for signing jwt token I've
abstracted out a new auth struct called Keypair which is used to
represent the underyling public and secret key used to sign data.

This struct is now used also by the Credentials struct as a part of our
login flow.