Hook up auth.Keypair with controller.Config
As a pre-requisite to enabling sessions to live between controller
restarts we needed to add a way to derive keypairs determinisitcally
which we can then use inside the token authority!

This commit introduces all of the code minus actually hooking up the
command line tooling and local development environment workflow to
enable sessions to live beyond a single restart.

Related dropoutlabs/planning#901