Log which component and it's port being listened on
Previously, the `framework.Server` was not logging out which port it was
listening on. This meant that we had no idea if the server was actually
up and working!

Now, the server requires a logger and uses the InstanceID from the
components config to figure out what's component it's running!