Adds the ability to attach and detach policy
To implement this had to create a new graphql route called policyByLabel
that allows us to query a policy given a label. Also had to modify the
attach route to make it so you could pass a `PolicyInput` object.
Unfortunately you can't pass our normal `Policy` object as mutation
routes only accept scalars, enums and other input objects. ????????

Also snuck in a policies list command for debugging CLI. We should
extend this to display attachments in the future.