Move GraphQL & Schema packages
These packages are a part of the controller and used by things outside
of the controller.

To that end, I've moved them into the Controller and separated out the
schemas from where the resolves and other code lives.

Small steps towards organizing the structure a little more now that we
have a bunch of different services!