Assign system roles to Users & Services
Previously, users and new services were not assigned to the appropriate
system roles (e.g. global). This commit introduces this behaviour.

At the same time, I've also introduced the ability to query roles for a
user and pull back a user's roles through GraphQL. This was needed to do
integration testing on the role assignment.

While I was at it, I also did a bit of refactoring to clean up a few of
the code routes.

Closes capeprivacy/cape#942