Fix bugs where EnvVar/Arg name were changed
Previously, if you changed the name of a EnvVar or Argument but forgot
to update the different commands you'd encounter panics at runtime.
Since we don't have great testing coverage for this section of the code
base this often meant breaking changes were getting into master.

To fix this, I've updated the code to use the reference to the argument
or env var. This will force this category of bug to be compiler errors.