Fix bug preventing login if no auth token set
Previously, if you had a cluster configured (via `config clusters add`)
and tried to login then we'd propagate a non-nil (but empty) auth token
to the client which it would then use to send an Authorization header.

This commit fixes this bug by checking for this edge case in the client
and preventing it inside the config package.