Prevent bad sql for an empty In block
This commit introduces a fix for fixing the bug where an empty in
statemnet would result in an sql error being propagated back to the
user.

We now check if the `in` is empty before we build up our query
statement. If you are looking for an exact field via the `QueryOne`
command then we return an error.

If you are looking for many results via the `Query` command then we just
treat the query as a no-op retaining the same behaviour as if you were
to query for something that doesn't exist.

Closes capeprivacy/planning#973