Make CLI testable
Closes https://github.com/capeprivacy/planning/issues/877

This adds a few interfaces that make it easier to test the CLI flow

Providers are registered as part of the `Before` middleware of the application.
The added provider can give you a UI to interact with, or a coordinator
client to make requests with. We can potentially add more things it provides
as needed. This is a pretty lightweight dependency injection flow.

The client now uses a transport, which is also obtained by the provider. The transport
handles making the actual request, as well as authorization.

For the CLI tests, we overwrite the `Before` middleware of the app and use a test provider
which will give a mocked transport.  See sources tests for an example