Fix Coordinator Signal, Add Signal to Connector
This commit introduces a new framework component: the `SignalWatcher`
which is a struct that watches for os signals and then attempts to shut
down a running service.

If the service doesn't shutdown cleanly in the configured time period it
will then hard crash the server and return an os exitcode of 1.

These changes also fix the bug preventing the coordinator from shutting
down cleanly when it receives a signal!

Related capeprivacy/planning#993