Introduce KinD
This commit introduces `KinD` to the Mage workflow enabling a user to
create and destroy a local kubernetes cluster without having to use k8s
run by Docker for Mac or other environments.

This is only part of the story, as in, coming commits I will need to
setup port forwarding and an in-cluster docker registry. The latter is
required so we can push docker images from the host into the cluster.

Inside the cluster, we can then use FluxCD to watch the local docker
registry to redeploy our helm charts! Boom, very similar workflow
locally as we would have in production!

Related capeprivacy/planning#1023