Add some edge case authorization decisions
Closes https://github.com/capeprivacy/planning/issues/1074

This is a workaround so that we can close off this authz work without
implementing a DSL just yet.

There are still a couple edge cases left around creating tokens. A user
can only create a token for themselves and a admin is the only role that
can create tokens for data connectors.