Condense dependency declaration for mage
Since every dependency is required and would panic if not present, this
introduces a `MustX()` constructor for each dependency type. Since these
constructors are all single return, it's easy to put the constructed
types in a list to add to the deps struct in a loop. This should make it
easy to avoid an accidental mismatch between dependency versions being
declared and those versions actually being registered.