Switch to Argon2ID w/ Pre-Shared Key
This commit switches our authentication flow from a two-phase Secure
Remote Password style-protocol to a single-phase pre-shared key
workflow.

To hash the passwords on the server, we're using Argon2ID which won the
Password Hashing Challenge.

For testing, because Argon2ID is memory and time intensive, we are using
SHA256 which is a _not_ a key derivation function. However, this change
_drastically_ expedites the speed of the tests across the board as we no
longer need to perform a computationally intensive key derivation in the
middle of our loops.

This change retains the concept of a CredentialProvider and continues to
store the Credentials object on the User & Token structs. The
credentials property on both structs is encrypted-at-rest inside of the
database using our `kms` functionality. We also strip the `Credentials`
from both structs before ever sending them over the wire.

We never want to leak these values out of the database in any capacity.

Due to the nature of the change, there are some drastic changes to flows
that previously relied on the client to provide all of the secret
material.

- When creating a user the password is not generated by the server and
  hashed before ever being sent to the user. In future, the invitation
  flow will remove this from occuring at all as we will simply generate
  a one-time-token that expires at some future date instead of
  generating a password at all.
- When creating a token, the secret is generated on the server and
  hashed before being sent back over the wire to be packaged up.
- We no longer need to create a "fake identity" as there is no
  enumeration vulnerability.

In addition to these changes, I've also introduced the following
functionality & concepts:

- There is now a `CredentialFactory` which is used for generating &
  comparing credentials based on their algorithm.

When this commit is merged, we can close up any work related to
improving the two-phase algorithm. We will need to open up new issues
for upgrading the Python library. Long term, it will be easier for us to
maintain this flow & build out client libraries.