Remove CredentialFactory
This pull request quickly removes the CredentialFactory removing support
for any form of backwards compatibility between comparisons. When we
need that behaviour it'll be pretty easy for us to layer on-top of what
we have.

At the same time, I've added some more test coverage to the
CredentialProducer's we had already defined!