Prompt if unlinking a service
When running `cape sources update`, failing to specify the
`--set-data-connector` flag effectively unlinks any data connector
from the source. This could be surprising behavior if, for example, the
flag was omitted in error. This change adds a prompt to confirm
unlinking the service, as well as a flag to automatically accept the
change.