Remove URL from APIToken
This commit removes the `url` property from the APIToken and introduces
a `CAPE_COORDINATOR_URL` environment variable enabling the coordinator
url to be configured independent of the token.

We are making this change as the internal address of a coordinator that
a connector or worker would use could be different than the external
address used by a CLI or UI request. This also enables a user to change
the DNS of a coordinator without needing to re-issue new identities to
the services.