Introduce ReAuthTransport
This commit introduces the ReAuthTransport which attempts to
reauthenticate with Cape in the case that their session expires.

The transport only authenticates APITokens as it's designed to
exclusively be used by our Cape services.

As a part of implementing the ReAuthTransport I made changes to the
relationship between a transport and a client. To remove duplication,
I've introduced a few private helper functions that leverage this
interface.