Integrate ReAuthTransport w/ Connector & Worker
This commit integrates the ReAuthTransport with the Connector & Worker
removing some pre-existing code that existed for managing authenticated
sessions.

The worker previously relied on the coordinator to be up for it to
start. These two services are no longer bound to one another. Worker
will keep on working and will just fail its jobs.