Only destroy registry if explicitly told
Previously, we destroyed the registry every time we threw the cluster
away. However, we don't need to do this as the docker registry lifecycle
& cluster lifecycle are different.

This will make it faster to destroy a cluster and spin it back up as
much of the docker context won't need to be loaded into the registry
from scratch.