Fixes bug where args from yaml not the right types
`Args` is currently mapped to `map[string]interface{}` when the
yaml/json decoder encounters a JSON number to be decoded to a
`interface{}` it decodes it as a float64. In the transformations args we
need to read them as float64s and then convert them into whatever form
of int we need.