Add support for paths to update command
This makes the `update` command take an optional list of paths to search
for database migrations to run on the configured db. This will allow us
to bundle migrations for the databases associated with all supported
services into a single container image or binary, and select the
appropriate migrations at run time.