The coordinators setups the database on first start
This replaces the need to have a separate command that sets up the
coordinators. The coordinator's configuration now takes a user config
which is required on first start and ignored otherwise. The setup
happens in the constructor of the coordinator and conditionally happens
depending on if the setup row is in the config table yet.