This adds secret args to named transformations
When a named transform in a project spec has a secret arg this arg is
converted in the SecretArg type. When this is submitted to the database
the secret value is randomly generated and then encrypted. When the
project spec with this named transform is requested then the secret
value is decrypted and delivered.